Add access Token and gets stars for python.
The current version pulls the stars for 100 repositories that use python and gets the average star count for all of them. The Idea I have for my project is a graph which displays the average star rating of repositories in different languages and organizations. This in my view will give a rough estimate to on average which language has the best programmers and which organizations have the best programmers.