Generates points for graph.
The program now makes API calls and calculates the data I will be using for my graph. I am getting the adverage star rating for google employees currently and the number of languages they have used in their repos. The project is still scaled down for testing purposes however This should fulfill part 1 of the assignment.