Uncommented_finished
This is the current commented finished version of the process. The process takes the data from 5 organizations and produces 2 graphs. 1 a scatter plot which plots each member of the organization in terms of their average star rating of their repositories and the amount of programming languages they have used. Graph 2 shows the language with the highest scoring repositories in the company and shows how it compares with the other company'e best language.